requirements.txt
pygame
sounddevice
soundfile
python_speech_features
scikit-learn
numpy
scipy
openai-whisper